// src/data/types/asistente.ts
export interface Asistente {
  IdAsistente: number
  Nombre: string | null
  DNI: string | null
  CUIT: string | null
  Genero: string | null
  Email: string | null
  DomicilioElectronico: string | null
  Rol: string | null
  Caracter: string | null
  EsAbogado: boolean | null
  IdAudiencia: number | null
  ValidacionRenaper: string | null
  ImagenValidacion: string | null
  Presente: boolean | null
  TipoAcceso: string | null
  Origen: string | null
  VersionColumnName: string | null
}